requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
chromedriver-autoinstaller==0.6.2
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
h11==0.14.0
idna==3.4
mysqlclient==2.2.0
numpy==1.26.0
outcome==1.2.0
pandas==2.1.1
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.6
wsproto==1.2.0
